requirements.txt
requests==2.31.0
google-generativeai==0.3.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
